Refactorizar otros componentes globales
